+++ NEWS.md
@@ -1,8 +1,27 @@
 # bbricks 0.1.0 (2020-02-28)
 This is the first release of bbricks.
+
 ## New features
-None
+**Model Structures:**
++ Gaussian and NIW conjugate. 
++ Gaussian and NIG conjugate.
++ Categorical and Dirichlet conjugate.
++ Categorical and Dirichlet process.
++ Categorical and 
++ Dirichlet process.
++ Hierarchical Dirichlet process.
++ Hierarchical Dirichlet process with two layers of hierarchies.
+
+**Tasks:**
++ Update posterior info.
++ Calculate sample sufficient statistics.
++ Calculate MAP estimate.
++ Calculate marginal likelihood.
++ Calculate posterior predictive density.
++ Sample from posterior predictive distribution.
+
 ## Improvements
 None
+
 ## Bug fixes
 None